pushing up work in progress..

+++ src/main/java/gov/nysenate/openleg/dao/sobi/SqlSOBIFileDao.java
@@ -8,18 +8,17 @@ import org.apache.commons.io.FileUtils;
 import org.apache.log4j.Logger;
 import org.joda.time.LocalDate;
 import org.springframework.dao.DataAccessException;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.dao.IncorrectResultSizeDataAccessException;
+import org.springframework.dao.DuplicateKeyException;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.*;
 
-import static gov.nysenate.openleg.util.FileHelper.moveFileToDirectory;
 import static gov.nysenate.openleg.util.FileHelper.safeListFiles;
 
 /**
@@ -65,12 +64,14 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
         "SELECT * FROM " + table(SOBI_FILE_TABLE) + " WHERE pending_processing = true ";
 
     private final String INSERT_SOBI_FILE =
-        "INSERT INTO " + table(SOBI_FILE_TABLE) + " (file_name, published_date_time, processed_date_time, processed_count, pending_processing) " +
-        "VALUES (:fileName, :publishedDateTime, :processedDateTime, :processedCount, :pendingProcessing)";
+        "INSERT INTO " + table(SOBI_FILE_TABLE) + "\n" +
+        "(file_name, published_date_time, staged_date_time, processed_date_time, processed_count, pending_processing) " +
+        "VALUES (:fileName, :publishedDateTime, :stagedDateTime, :processedDateTime, :processedCount, :pendingProcessing)";
 
     private final String UPDATE_SOBI_FILE =
         "UPDATE " + table(SOBI_FILE_TABLE) + "\n" +
         "SET published_date_time = :publishedDateTime," +
+        "    staged_date_time = :stagedDateTime," +
         "    processed_date_time = :processedDateTime," +
         "    pending_processing = :pendingProcessing," +
         "    processed_count = :processedCount " +
@@ -83,19 +84,7 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     public SOBIFile getSOBIFile(String fileName) {
         SOBIFile sobiFile = null;
         MapSqlParameterSource params = new MapSqlParameterSource("fileNames", Arrays.asList(fileName));
-        try {
-            SOBIMetadata metaData = jdbcNamed.queryForObject(GET_SOBI_FILES_BY_FILE_NAMES_SQL, params, new SOBIMetadata.Mapper());
-            sobiFile = new SOBIFile(getSOBIFileFromFileSystem(metaData));
-            metaData.applyToSOBIFile(sobiFile);
-        }
-        catch (IncorrectResultSizeDataAccessException ex) {
-            logger.debug("No sobi details were found in database for " + fileName, ex);
-        }
-        catch (IOException ex) {
-            logger.error("Possible inconsistency between file system and database! " +
-                         "Failed to construct SOBIFile instance with file name " + fileName, ex);
-        }
-        return sobiFile;
+        return jdbcNamed.queryForObject(GET_SOBI_FILES_BY_FILE_NAMES_SQL, params, new SOBIFileRowMapper());
     }
 
     /** {@inheritDoc }*/
@@ -103,17 +92,9 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     public Map<String, SOBIFile> getSOBIFiles(List<String> fileNames) {
         MapSqlParameterSource params = new MapSqlParameterSource("fileNames", fileNames);
         Map<String, SOBIFile> sobiFileMap = new HashMap<>();
-        List<SOBIMetadata> metaDataList = jdbcNamed.query(GET_SOBI_FILES_BY_FILE_NAMES_SQL, params, new SOBIMetadata.Mapper());
-        try {
-            for (SOBIMetadata metaData : metaDataList) {
-                SOBIFile sobiFile = new SOBIFile(getSOBIFileFromFileSystem(metaData));
-                metaData.applyToSOBIFile(sobiFile);
-                sobiFileMap.put(sobiFile.getFileName(), sobiFile);
-            }
-        }
-        catch (IOException ex) {
-            logger.error("Failed to construct SOBIFile instance!", ex);
-            return null;
+        List<SOBIFile> sobiList = jdbcNamed.query(GET_SOBI_FILES_BY_FILE_NAMES_SQL, params, new SOBIFileRowMapper());
+        for (SOBIFile sobiFile : sobiList) {
+            sobiFileMap.put(sobiFile.getFileName(), sobiFile);
         }
         return sobiFileMap;
     }
@@ -125,17 +106,14 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
         params.addValue("startDate", start);
         params.addValue("endDate", end);
         params.addValue("processedOnly", processedOnly);
-        List<SOBIMetadata> metaDataList = jdbcNamed.query(GET_SOBI_FILES_DURING_SQL + orderBy("published_date_time", sortByPubDate),
-                                                          params, new SOBIMetadata.Mapper());
-        return getSobiFileListFromMetaDataList(metaDataList);
+        return jdbcNamed.query(GET_SOBI_FILES_DURING_SQL + orderBy("published_date_time", sortByPubDate),
+                params, new SOBIFileRowMapper());
     }
 
     /** {@inheritDoc }*/
     @Override
     public List<SOBIFile> getPendingSOBIFiles(SortOrder sortByPubDate) {
-        List<SOBIMetadata> metaDataList =
-            jdbc.query(GET_PENDING_SOBI_FILES_SQL + orderBy("published_date_time", sortByPubDate), new SOBIMetadata.Mapper());
-        return getSobiFileListFromMetaDataList(metaDataList);
+        return jdbc.query(GET_PENDING_SOBI_FILES_SQL + orderBy("published_date_time", sortByPubDate),                new SOBIFileRowMapper());
     }
 
     /** {@inheritDoc }*/
@@ -143,45 +121,36 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     public void stageSOBIFiles(boolean allowReStaging) throws IOException {
         logger.info("Performing staging of new SOBI files...");
         int filesStaged = 0;
-
         for (File file : safeListFiles(this.stagingSobiDir, null, false)) {
             SOBIFile sobiFile = new SOBIFile(file);
+            sobiFile.setStagedDateTime(new Date());
             sobiFile.setPendingProcessing(true);
             try {
                 insertSOBIMetadataInDatabase(sobiFile);
-                if (!file.delete()) {
-                    logger.fatal("Failed to delete " + file.getAbsoluteFile() + " from the sobi staging directory!");
-                }
-                updateSOBIFileInFileSystem(sobiFile);
+                moveFileToWorking(file);
                 filesStaged++;
                 logger.debug("Staged " + sobiFile);
             }
-            catch(DataIntegrityViolationException ex) {
+            catch (DuplicateKeyException ex) {
                 if (allowReStaging) {
                     // The SOBI file has been staged once before so we want to re-stage it with the contents of the
                     // new file which may or may not be identical. This should rarely occur in production but can
                     // be helpful during development.
                     SOBIFile existingSOBIFile = getSOBIFile(sobiFile.getFileName());
                     if (existingSOBIFile != null) {
-                        File existingFile = getSOBIFileFromFileSystem(existingSOBIFile);
-                        if (existingFile.exists()) {
-                            moveFileToDirectory(file, this.workingSobiDir, true);
-                            existingSOBIFile.setPendingProcessing(true);
-                            updateSOBIFile(existingSOBIFile);
-                            filesStaged++;
-                            logger.debug("Re-staged " + existingSOBIFile + " with latest file contents.");
-                        }
-                        else {
-                            logger.warn("Attempted to re-stage " + existingSOBIFile + " but the file " +
-                                    existingFile.getAbsolutePath() + " doesn't exist!");
-                        }
+                        existingSOBIFile.setStagedDateTime(new Date());
+                        existingSOBIFile.setPendingProcessing(true);
+                        updateSOBIFile(existingSOBIFile);
+                        moveFileToWorking(file);
+                        filesStaged++;
+                        logger.debug("Re-staged " + existingSOBIFile + " with latest file contents.");
                     }
                     else {
                         logger.warn("Failed to stage " + sobiFile + " because a prior record exists but cannot be retrieved!");
                     }
                 }
                 else {
-                    logger.debug("Skipping " + sobiFile + " since re-staging is disabled.");
+                    logger.debug("Ignoring " + sobiFile + " since re-staging is disabled and file was previously processed.");
                 }
             }
         }
@@ -193,9 +162,14 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     public boolean updateSOBIFile(SOBIFile sobiFile) {
         if (sobiFile != null) {
             try {
-                int affected = jdbcNamed.update(UPDATE_SOBI_FILE, getSOBIInsertUpdateParams(sobiFile));
+                int affected = updateSOBIMetadataInDatabase(sobiFile);
                 if (affected > 0) {
-                    updateSOBIFileInFileSystem(sobiFile);
+                    if (!sobiFile.isPendingProcessing()) {
+                        File workFile = getFileInWorkingDirectory(sobiFile.getFileName());
+                        if (workFile.exists()) {
+                            moveFiletoArchive(workFile, sobiFile.getPublishedDateTime());
+                        }
+                    }
                     return true;
                 }
                 return false;
@@ -226,47 +200,58 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
 
     /** --- Helper Classes --- */
 
-    /** Models the processing details for a sobi file which is stored in the database */
-    protected static class SOBIMetadata {
-        String fileName;
-        Date publishedDateTime;
-        Date processedDateTime;
-        boolean pendingProcessing;
-        int processedCount;
+    /**
+     * Maps rows from the sobi table to SOBIFile objects.
+     */
+    protected class SOBIFileRowMapper implements RowMapper<SOBIFile> {
+        private String pfx = "";
 
-        /** Apply metadata info to SOBIFile */
-        void applyToSOBIFile(SOBIFile sobiFile) {
-            sobiFile.setProcessedDateTime(this.processedDateTime);
-            sobiFile.setPendingProcessing(this.pendingProcessing);
-            sobiFile.setProcessedCount(this.processedCount);
+        public SOBIFileRowMapper(){}
+        public SOBIFileRowMapper(String pfx) {
+            this.pfx = pfx;
         }
 
-        /** Row Mapper implementation */
-        static class Mapper implements RowMapper<SOBIMetadata> {
-            @Override
-            public SOBIMetadata mapRow(ResultSet rs, int rowNum) throws SQLException {
-                SOBIMetadata metadata = new SOBIMetadata();
-                metadata.fileName = rs.getString("file_name");
-                metadata.publishedDateTime = rs.getDate("published_date_time");
-                metadata.processedDateTime = rs.getDate("processed_date_time");
-                metadata.pendingProcessing = rs.getBoolean("pending_processing");
-                metadata.processedCount = rs.getInt("processed_count");
-                return metadata;
+        @Override
+        public SOBIFile mapRow(ResultSet rs, int rowNum) throws SQLException {
+            String fileName = rs.getString(pfx + "file_name");
+            Date publishedDateTime = rs.getDate(pfx + "published_date_time");
+            boolean pendingProcessing = rs.getBoolean(pfx + "pending_processing");
+            File file = getSOBIFileFromFileSystem(fileName, publishedDateTime, pendingProcessing);
+            try {
+                SOBIFile sobiFile = new SOBIFile(file);
+                sobiFile.setPendingProcessing(pendingProcessing);
+                sobiFile.setProcessedDateTime(rs.getDate(pfx + "processed_date_time"));
+                sobiFile.setProcessedCount(rs.getInt(pfx + "processed_count"));
+                return sobiFile;
             }
+            catch (FileNotFoundException ex) {
+                logger.error("SOBI file " + rs.getString("file_name") + " was not found in the expected location! \n" +
+                        "This could be a result of modifications to the sobi file directory that were not synced with " +
+                        "the database. ", ex);
+            }
+            catch (IOException ex) {
+                logger.error(ex);
+            }
+            return null;
         }
     }
+
+    /** --- Internal Methods --- */
 
     /**
      * Saves the metadata of the SOBIFile into the database.
      * @throws java.lang.IllegalArgumentException if sobiFile is null
      */
     private void insertSOBIMetadataInDatabase(SOBIFile sobiFile) throws DataAccessException {
-        if (sobiFile != null) {
-            jdbcNamed.update(INSERT_SOBI_FILE, getSOBIInsertUpdateParams(sobiFile));
-        }
-        else {
-            throw new IllegalArgumentException("Supplied SOBIFile object is null. Cannot persist to database.");
-        }
+        jdbcNamed.update(INSERT_SOBI_FILE, getSOBIInsertUpdateParams(sobiFile));
+    }
+
+    /**
+     * Updates the metadata of the SOBIFile into the database.
+     * @throws java.lang.IllegalArgumentException if sobiFile is null
+     */
+    private int updateSOBIMetadataInDatabase(SOBIFile sobiFile) throws DataAccessException {
+        return jdbcNamed.update(UPDATE_SOBI_FILE, getSOBIInsertUpdateParams(sobiFile));
     }
 
     /**
@@ -276,13 +261,8 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
      * @param pendingProcessing boolean - Indicates where to look for the SOBI file.
      * @return File
      */
-    private File getSOBIFileFromFileSystem(String fileName, Date publishedDate, boolean pendingProcessing) {
-        return (pendingProcessing) ? getFileInWorkingDirectory(fileName) : getFileInArchiveDirectory(fileName, publishedDate);
-    }
-
-    /** Overload for convenience. @see #getSOBIFileFromFileSystem(String, java.util.Date, boolean) */
-    private File getSOBIFileFromFileSystem(SOBIMetadata metaData) {
-        return getSOBIFileFromFileSystem(metaData.fileName, metaData.publishedDateTime, metaData.pendingProcessing);
+    private File getSOBIFileFromFileSystem(String fileName, Date publishedDateTime, boolean pendingProcessing) {
+        return (pendingProcessing) ? getFileInWorkingDirectory(fileName) : getFileInArchiveDirectory(fileName, publishedDateTime);
     }
 
     /** Overload for convenience. @see #getSOBIFileFromFileSystem(String, java.util.Date, boolean) */
@@ -291,30 +271,37 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     }
 
     /**
-     * Deletes the actual SOBI file from the file system and writes the contents of SOBIFile to the appropriate directory.
-     * @param sobiFile SOBIFile
-     * @return File - Newly written file.
-     * @throws IOException
+     * Moves the file into the working SOBI file directory.
+     * @param file File
      */
-    private File updateSOBIFileInFileSystem(SOBIFile sobiFile) throws IOException {
-        File workFile = getFileInWorkingDirectory(sobiFile.getFileName());
-        File archiveFile = getFileInArchiveDirectory(sobiFile.getFileName(), sobiFile.getPublishedDateTime());
-
-        // Delete the file from any possible directory so we don't have to keep track of where it is.
-        FileUtils.deleteQuietly(workFile);
-        FileUtils.deleteQuietly(archiveFile);
-
-        if (sobiFile.isPendingProcessing()) {
-            // Write to the working directory.
-            FileUtils.write(workFile, sobiFile.getText());
-            return workFile;
+    private void moveFileToWorking(File file) throws IOException {
+        File workingFile = getFileInWorkingDirectory(file.getName());
+        if (workingFile.exists()) {
+            FileUtils.deleteQuietly(workingFile);
         }
-        else {
-            // Create any sub directories just to be safe and write to archive directory.
-            archiveFile.getParentFile().mkdirs();
-            FileUtils.write(archiveFile, sobiFile.getText());
-            return archiveFile;
+        FileUtils.moveFile(file, workingFile);
+    }
+
+    /**
+     * Moves the file into the archive SOBI file directory.
+     * @param file File
+     */
+    private void moveFiletoArchive(File file, Date publishedDateTime) throws IOException {
+        File archiveFile = getFileInArchiveDirectory(file.getName(), publishedDateTime);
+        if (archiveFile.exists()) {
+            FileUtils.deleteQuietly(archiveFile);
         }
+        FileUtils.moveFile(file, archiveFile);
+    }
+
+    /**
+     * Delete any traces of the SOBI file (either in work or archive directory) from the file system.
+     * @param fileName String
+     * @param publishedDateTime Date
+     */
+    private void deleteSOBIFileFromFileSystem(String fileName, Date publishedDateTime) {
+        FileUtils.deleteQuietly(getFileInWorkingDirectory(fileName));
+        FileUtils.deleteQuietly(getFileInArchiveDirectory(fileName, publishedDateTime));
     }
 
     /**
@@ -329,39 +316,17 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
     /**
      * Get file handle from the sobi archive directory.
      * @param fileName String
-     * @param publishedDate Date
+     * @param publishedDateTime Date
      * @return File
      */
-    private File getFileInArchiveDirectory(String fileName, Date publishedDate) {
-        LocalDate localDate = new LocalDate(publishedDate);
-        String year = Integer.toString(localDate.getYear());
+    private File getFileInArchiveDirectory(String fileName, Date publishedDateTime) {
+        String year = Integer.toString(new LocalDate(publishedDateTime).getYear());
         File dir = new File(this.archiveSobiDir, year);
         return new File(dir, fileName);
     }
 
     /**
-     * Convenience method to convert a list of SOBIMetaData objects into SOBIFile objects.
-     * @param metaDataList List<SOBIMetaData>
-     * @return List<SOBIFile>
-     */
-    private List<SOBIFile> getSobiFileListFromMetaDataList(List<SOBIMetadata> metaDataList) {
-        List<SOBIFile> sobiFiles = new ArrayList<>();
-        try {
-            for (SOBIMetadata metaData : metaDataList) {
-                SOBIFile sobiFile = new SOBIFile(getSOBIFileFromFileSystem(metaData));
-                metaData.applyToSOBIFile(sobiFile);
-                sobiFiles.add(sobiFile);
-            }
-        }
-        catch (IOException ex) {
-            logger.error("Failed to construct SOBIFile instance!", ex);
-            return null;
-        }
-        return sobiFiles;
-    }
-
-    /**
-     * Returns a ParameterSource with columns mapped to SOBIFile values for use in update/insert queries.
+     * Returns a MapSqlParameterSource with columns mapped to SOBIFile values for use in update/insert queries.
      * @param sobiFile SOBIFile
      * @return MapSqlParameterSource
      */
@@ -370,6 +335,7 @@ public class SqlSOBIFileDao extends SqlBaseDao implements SOBIFileDao
         params.addValue("fileName", sobiFile.getFileName());
         params.addValue("publishedDateTime", toTimestamp(sobiFile.getPublishedDateTime()));
         params.addValue("processedDateTime", toTimestamp(sobiFile.getProcessedDateTime()));
+        params.addValue("stagedDateTime", toTimestamp(sobiFile.getStagedDateTime()));
         params.addValue("processedCount", sobiFile.getProcessedCount());
         params.addValue("pendingProcessing", sobiFile.isPendingProcessing());
         return params;
